Create separate inertia apps
